pyproject.toml
[project]
name = "graphql-api"
version = "0.1.0"
description = "API GraphQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker",
    "fastapi[standard]",
    "requests",
    "strawberry-graphql[fastapi]",
    "uvicorn",
    "pymongo",
]

[project.optional-dependencies]
test = ["httpx", "pytest"]

[tool.setuptools.packages.find]
where = ["src"]
